ExampleDO生成

+++ apocalypse-example/apocalypse-example-app/src/main/java/com/apocalypse/example/mapper/ExampleDOMapper.java
@@ -3,5 +3,5 @@ package com.apocalypse.example.mapper;
 import com.apocalypse.common.mybatis.MyMapper;
 import com.apocalypse.example.model.ExampleDO;
 
-public interface ExampleDOMapper extends MyMapper<ExampleDO, Integer> {
+public interface ExampleDOMapper extends MyMapper<ExampleDO, Long> {
 }
